--- rollbar-java/src/main/java/com/rollbar/notifier/exception/SenderException.java
@@ -1,13 +1,0 @@
-package com.rollbar.notifier.exception;
-
-import com.rollbar.notifier.sender.Sender;
-
-/**
- * Exception to indicate that there was a problem related in the {@link Sender sender}.
- */
-public class SenderException extends RuntimeException {
-
-  public SenderException(Exception e) {
-    super(e);
-  }
-}
