+++ rollbar-java/src/main/java/com/rollbar/notifier/sender/AbstractSender.java
@@ -7,6 +7,8 @@ import com.rollbar.notifier.sender.exception.ApiException;
 import com.rollbar.notifier.sender.exception.SenderException;
 import com.rollbar.notifier.sender.listener.SenderListener;
 import com.rollbar.notifier.sender.result.Response;
+
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import org.slf4j.Logger;
@@ -56,7 +58,7 @@ public abstract class AbstractSender implements Sender {
   }
 
   @Override
-  public void close() throws Exception {
+  public void close() throws IOException {
 
   }
