+++ rollbar-java/src/main/java/com/rollbar/notifier/sender/BufferedSender.java
@@ -5,6 +5,8 @@ import com.rollbar.notifier.config.Config;
 import com.rollbar.notifier.sender.exception.SenderException;
 import com.rollbar.notifier.sender.listener.SenderListener;
 import com.rollbar.notifier.sender.queue.DiskQueue;
+
+import java.io.IOException;
 import java.util.List;
 import java.util.Objects;
 import java.util.Queue;
@@ -37,8 +39,8 @@ public class BufferedSender implements Sender {
   }
 
   BufferedSender(Builder builder, ScheduledExecutorService executorService) {
-    Objects.requireNonNull(builder.sender, "The sender can not be null");
-    Objects.requireNonNull(builder.queue, "The queue can not be null");
+    requireNonNull(builder.sender, "The sender can not be null");
+    requireNonNull(builder.queue, "The queue can not be null");
 
     this.batchSize = builder.batchSize;
     this.sender = builder.sender;
@@ -49,6 +51,12 @@ public class BufferedSender implements Sender {
     this.executorService =  executorService;
     this.executorService.scheduleWithFixedDelay(new SendTask(batchSize, queue, sender),
         builder.initialFlushDelay, builder.flushFreq, TimeUnit.MILLISECONDS);
+  }
+  
+  private <T> T requireNonNull(T obj, String message) {
+    if (obj == null)
+      throw new NullPointerException(message);
+    return obj;
   }
 
   @Override
@@ -72,7 +80,7 @@ public class BufferedSender implements Sender {
   }
 
   @Override
-  public void close() throws Exception {
+  public void close() throws IOException {
     this.executorService.shutdown();
     this.sender.close();
   }
